etcdserver: move v3 raft apply functions to interface
